+++ src/config/activemenu.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+
+    // The default css class value if the request match given route name
+    'class' => 'active',
+
+];